Create google-oauth-2-login-button
This is a button which allows submission of oauth requests with a local server middleman to allow login to google.  It uses an Async JS pattern to load all the information to the page you're on, including the access_token and profile.  Server side portion must be done in a server side language and must hanlde callback url, access_token, and sending profile.

Oauth 2 is essentially a three step process of first hitting the auth URL and logging in, then verifying your client again with an encrypted code and only then do you get the profile back.

This is a long way of saying this is only half the answer to the problem.  The other half I've solved in node and php by making a very simple API with a session on a third party page which returns the code, then the access token and profile after logging in